Source out Actions helper structs
