+++ source/Leviathan/input/ActionMapping.cpp
@@ -1,0 +1,11 @@
+#include "ActionMapping.h"
+
+namespace leviathan {
+    namespace input {
+        ActionMapping::ActionMapping(const YAML::Node& node)
+        : name(node && node["name"] ? node["name"].as<std::string>() : "nameless action"),
+          description(node && node["description"] ? node["description"].as<std::string>() : ""),
+          id(node && node["id"] ? node["id"].as<uint32_t>() : 0), internal(node && node["internal"]),
+          primary(node["input_mappings"]["primary"]), secondary(node["input_mappings"]["secondary"]) {}
+    }
+}
