+++ source/Leviathan/input/EventToActionConverter.h
@@ -7,8 +7,8 @@
 #ifndef LEVIATHAN_INPUT_EVENT_TO_ACTION_CONVERTER_H
 #define LEVIATHAN_INPUT_EVENT_TO_ACTION_CONVERTER_H
 
+#include "Action.h"
 #include "irrlicht.h"
-#include "types.h"
 #include <list>
 #include <map>
 #include <vector>
